rpi/requirements.txt
mfrc522-python
requests
RPLCD
spidev
RPi.GPIO
